% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal_median.R
\name{cal_median}
\alias{cal_median}
\title{Title}
\usage{
cal_median(data, colName)
}
\arguments{
\item{data}{dataframe}

\item{colName}{column name}
}
\value{
calMedian median of the column
}
\description{
Title
}
\examples{
nobleGasses <- data.frame(Gas=c("Helium","Neon", "Argon", "Krypton", "Xenon", "Radon", "Oganesson"),
Symbol=c("He","Ne","Ae","Kr","Xe","Rn","Og"), Nr=c(2,10,18,36,54,86,118),
Mass=c(4.003,20.18,39.948,83.8,131.293,222,294))
cal_median(nobleGasses,"Mass")  # returns 83.3
cal_median(nobleGasses,4)       # returns 83.3
}
