% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_NA.R
\name{check_NA}
\alias{check_NA}
\title{Title count NA in a data column}
\usage{
check_NA(data, colName)
}
\arguments{
\item{data}{dataframe}

\item{colName}{column name}
}
\value{
number of NA's in the column
}
\description{
Title count NA in a data column
}
\examples{
check_NA(sample_read(),"TotalCharges")
}
