% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_negative.R
\name{check_negative}
\alias{check_negative}
\title{Title check Negative values in a dataframe}
\usage{
check_negative(data)
}
\arguments{
\item{data}{dataframe}
}
\value{
number of Negative values in dataframe
}
\description{
Title check Negative values in a dataframe
}
\examples{
A = data.frame(B=c(1,2,3,-4),C=c(-1,-2,3,4))
check_negative(A)  # return 3
}
