% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_negative_col.R
\name{check_negative_col}
\alias{check_negative_col}
\title{Title check Negative values in a data column}
\usage{
check_negative_col(data, colName)
}
\arguments{
\item{data}{dataframe}

\item{colName}{column name}
}
\value{
number of Negative values in the column
}
\description{
Title check Negative values in a data column
}
\examples{
A = data.frame(B=c(1,2,3,-4),C=c(-1,-2,3,4))
check_negative_col(A,"B") #return 1
check_negative_col(A,"C") #return 2
}
