% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_normal_distribution.R
\name{check_normal_distribution}
\alias{check_normal_distribution}
\title{Title check if column distribution is normal}
\usage{
check_normal_distribution(data, colName)
}
\arguments{
\item{colName}{}
}
\value{
TRUE if column has normal distribution
}
\description{
Title check if column distribution is normal
}
\examples{
#' nobleGasses <- data.frame(Gas=c("Helium","Neon", "Argon", "Krypton", "Xenon", "Radon", "Oganesson"),
Symbol=c("He","Ne","Ae","Kr","Xe","Rn","Og"), Nr=c(2,10,18,36,54,86,118),
Mass=c(4.003,20.18,39.948,83.8,131.293,222,294))
check_normal_distribution(nobleGasses,"Mass")  # return FALSE
}
