% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chisq_test.R
\name{chisq_test}
\alias{chisq_test}
\title{Title compute chi square test for labelled column}
\usage{
chisq_test(data, colName)
}
\arguments{
\item{data}{dataframe}

\item{colName}{label column name}
}
\value{
chi square test parameters
}
\description{
Title compute chi square test for labelled column
}
\examples{
chisq_test(iris,"Species")    #return Chi-squared test for given probabilities data:  count_each_label X-squared = 0, df = 2, p-value = 1
}
