% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohens_d.R
\name{cohens_d}
\alias{cohens_d}
\title{Title Dependent t test}
\usage{
cohens_d(data, colName)
}
\arguments{
\item{data}{dataframe}

\item{colName}{columname}
}
\value{
cohen's d value
}
\description{
Title Dependent t test
}
