% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/col_compare_label.R
\name{col_compare_label}
\alias{col_compare_label}
\title{Title}
\usage{
col_compare_label(data, colName, colLabel)
}
\arguments{
\item{data}{dataframe}

\item{colName}{name of the column needed to check rate in relation to labeled column}

\item{colLabel}{name of labeled column}
}
\value{
a titble based on given and labeled column
}
\description{
Title
}
