% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/col_type.R
\name{col_type}
\alias{col_type}
\title{Use a map function to find which columns are numeric, the type of each column, and a summary of each column.}
\usage{
col_type(df)
}
\arguments{
\item{data}{A data frame or matrix}
}
\value{
Vector of colun type
}
\description{
Use a map function to find which columns are numeric, the type of each column, and a summary of each column.
}
\examples{
col_type(airquality)

}
