% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/column_std.R
\name{column_std}
\alias{column_std}
\title{Title calculate the standard deviation}
\usage{
column_std(data, colName)
}
\arguments{
\item{data}{dataframe}

\item{colName}{columname}
}
\value{
standard deviation of an observation variable
}
\description{
Title calculate the standard deviation
}
\examples{
nobleGasses <- data.frame(Gas=c("Helium","Neon", "Argon", "Krypton", "Xenon", "Radon", "Oganesson"),
Symbol=c("He","Ne","Ae","Kr","Xe","Rn","Og"), Nr=c(2,10,18,36,54,86,118),
Mass=c(4.003,20.18,39.948,83.8,131.293,222,294))
}
