% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/column_variance.R
\name{column_variance}
\alias{column_variance}
\title{Title calculate the variance}
\usage{
column_variance(data, colName)
}
\arguments{
\item{data}{dataframe}

\item{colName}{columname}
}
\value{
variance of an observation variable
}
\description{
Title calculate the variance
}
