% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_col_distribution_2data.R
\name{compare_col_distribution_2data}
\alias{compare_col_distribution_2data}
\title{Title check if column distribution of two dataframe are same}
\usage{
compare_col_distribution_2data(data1, data2, colName)
}
\arguments{
\item{data1}{data frame}

\item{data2}{data frame}

\item{colName}{column name}
}
\value{
TRUE if column has different distribution
}
\description{
Title check if column distribution of two dataframe are same
}
\examples{
nobleGasses <- data.frame(Gas=c("Helium","Neon", "Argon", "Krypton", "Xenon", "Radon", "Oganesson"),
Symbol=c("He","Ne","Ae","Kr","Xe","Rn","Og"), Nr=c(2,10,18,36,54,86,118),
Mass=c(4.003,20.18,39.948,83.8,131.293,222,294))
NnobleGasses <- data.frame(Gas=c("Helium","Neon", "Argon", "Krypton", "Xenon", "Radon", "Oganesson"),
Symbol=c("He","Ne","Ae","Kr","Xe","Rn","Og"), Nr=c(2,10,18,36,54,86,118),
Mass=c(4.003,20.18,39.948,83.8,131293,222,29400))
compare_col_distribution_2data(nobleGasses,hnobleGasses,"Nr")    #return FALSE
compare_col_distribution_2data(nobleGasses,hnobleGasses,"Mass")  # return TRUE
}
