% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_two_values.R
\name{compare_two_values}
\alias{compare_two_values}
\title{Title compare 2 columns values}
\usage{
compare_two_values(data, colName1, colName2)
}
\arguments{
\item{data}{data frame}

\item{colName1}{first column name}

\item{colName2}{second column name}
}
\value{
number of times column have different values then condition
}
\description{
Title compare 2 columns values
}
\examples{
nobleGasses <- data.frame(Gas=c("Helium","Neon", "Argon", "Krypton", "Xenon", "Radon", "Oganesson"),
Symbol=c("He","Ne","Ae","Kr","Xe","Rn","Og"), Nr=c(2,10,18,36,54,86,118),
Mass=c(4.003,20.18,39.948,83.8,131.293,222,294))
compare_two_values(nobleGasses, "Mass","Nr")   # reutrns 7
compare_two_values(nobleGasses,"Nr", "Mass")   # reutrns 0
}
