% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_data_variable.R
\name{convert_data_variable}
\alias{convert_data_variable}
\title{Title Convert Variables format}
\usage{
convert_data_variable(data)
}
\arguments{
\item{data}{dataframe read}
}
\value{
processed data
}
\description{
Title Convert Variables format
}
\examples{

}
