% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_column_range.R
\name{count_column_range}
\alias{count_column_range}
\title{Title check the column values are in the given range}
\usage{
count_column_range(data, colName, loVal, hiVal)
}
\arguments{
\item{data}{dataframe}

\item{colName}{column name to check in the range}

\item{loVal}{lower value of the range}

\item{hiVal}{higher value of the range}
}
\value{
total rows that are in range
}
\description{
Title check the column values are in the given range
}
\examples{
nobleGasses <- data.frame(Gas=c("Helium","Neon", "Argon", "Krypton", "Xenon", "Radon", "Oganesson"),
Symbol=c("He","Ne","Ae","Kr","Xe","Rn","Og"), Nr=c(2,10,18,36,54,86,118),
Mass=c(4.003,20.18,39.948,83.8,131.293,222,294))
check_column_range(nobleGasses, "Mass",4,50)   # reutrns 3
check_column_range(nobleGasses, "Mass",30,50)   # reutrns 1
}
