% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_summary.R
\name{data_summary}
\alias{data_summary}
\title{Data Summary for Numeric Columns}
\usage{
data_summary(x, na.rm = TRUE)
}
\arguments{
\item{x}{A data.frame containing at least one numeric column}

\item{na.rm}{A logical indicating whether missing values should be removed}
}
\value{
This function returns a \code{data.frame} including columns:
\itemize{
 \item ID
 \item min
 \item median
 \item sd
 \item max
}
}
\description{
Custom summaries of numeric data in a provided data frame
}
\examples{
data_summary(iris)
data_summary(airquality, na.rm = FALSE)


#'
}
\seealso{
\link[base]{summary}
}
\author{
Pushpinder Kaur Chouhan <p.chouhan@ulster.ac.uk>
}
