% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decision_tree_model.R
\name{decision_tree_model}
\alias{decision_tree_model}
\title{Title}
\usage{
decision_tree_model(data, colName, divRatio)
}
\arguments{
\item{data}{data frame}

\item{colName}{label column name}

\item{divRatio}{data split (test/train) ratio 75%, 85%}
}
\value{
modelTree decision tree model
}
\description{
Title
}
