% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependent_t_test.R
\name{dependent_t_test}
\alias{dependent_t_test}
\title{Title Dependent t test}
\usage{
dependent_t_test(data, colName)
}
\arguments{
\item{data}{dataframe}

\item{colName}{columname}
}
\value{
t-value
}
\description{
Title Dependent t test
}
