% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/divide_data_test_train.R
\name{divide_data_test_train}
\alias{divide_data_test_train}
\title{Title To split the dataset into training and test}
\usage{
divide_data_test_train(data, divRatio)
}
\arguments{
\item{data}{data frame}

\item{divRatio}{value for divinding the data into train and test}
}
\value{
list of train and test datasets
}
\description{
Title To split the dataset into training and test
}
\examples{
nobleGasses <- data.frame(Gas=c("Helium","Neon", "Argon", "Krypton", "Xenon", "Radon", "Oganesson"),
Symbol=c("He","Ne","Ae","Kr","Xe","Rn","Og"), Nr=c(2,10,18,36,54,86,118),
Mass=c(4.003,20.18,39.948,83.8,131.293,222,294))
divide_data_test_train(nobleGasses,75) # return list of 2 train and test dataframe
}
