% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_test_vector.R
\name{f_test_vector}
\alias{f_test_vector}
\title{Title check if 2 column (numeric vector) have the same variance}
\usage{
f_test_vector(data, colName1, colName2)
}
\arguments{
\item{data}{dataframe}

\item{colName1}{Name of the column}

\item{colName2}{Name of the column}
}
\value{
fTestVariance ftest console output as variable
}
\description{
Title check if 2 column (numeric vector) have the same variance
}
\examples{
dataTest <- data.frame(x=c(rnorm(30, mean = 0, sd = 2)), y=c(rnorm(30, mean = 1, sd = 1)))
f_test_vector(dataTest,"x","y")
}
