% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{hello}
\alias{hello}
\title{Customer Churn data analysis}
\usage{
hello(x)
}
\arguments{
\item{x}{}
}
\value{
confusion matrix of the model
}
\description{
Customer Churn data analysis
}
\examples{
hello("Akul")
\dontrun{
hello("Me")
}
}
