% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lg_Model.R
\name{lg_Model}
\alias{lg_Model}
\title{Title logistic regression model on the data}
\usage{
lg_Model(data, colName)
}
\arguments{
\item{divRatio}{}
}
\description{
Title logistic regression model on the data
}
\examples{
lg_Model(processed_data(sample_read()),churn,75)
}
