% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/look_data.R
\name{look_data}
\alias{look_data}
\title{view the uploaded dataset}
\usage{
look_data(data)
}
\arguments{
\item{data}{dataframe to look at}
}
\value{
confusion matrix of the model \code{\link{print}}
}
\description{
view the uploaded dataset
}
