% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_value.R
\name{match_value}
\alias{match_value}
\title{Title}
\usage{
match_value(data, colName, value1)
}
\arguments{
\item{data}{data frame}

\item{value1}{value to be checked for}

\item{colVal}{columnName or column Number that need to be checked for specific values}

\item{value2}{value to be chekced for}
}
\value{
total rows with matched value
}
\description{
Title
}
\examples{
from flight data check the delayed
iris$Species[which(match_value(iris,4,0.1,0.5))]
iris$Species[which(match_value(iris,"Petal.Width",0.1,0.5))]
}
