% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na_counter.R
\name{na_counter}
\alias{na_counter}
\title{Count all NAs in a data set}
\usage{
na_counter(data)
}
\arguments{
\item{data}{A data frame or matrix}
}
\value{
Vector of NA counts
}
\description{
Count all NAs in a data set
}
\examples{
na_counter(airquality)
}
