% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_col.R
\name{normalize_col}
\alias{normalize_col}
\title{Title Normalize numeric data}
\usage{
normalize_col(data, colName)
}
\arguments{
\item{data}{dataframe}

\item{colName}{column name}
}
\description{
Title Normalize numeric data
}
