% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omit_na.R
\name{omit_na}
\alias{omit_na}
\title{Title relatively small in number then ignore those rows from the analysis}
\usage{
omit_na(data)
}
\arguments{
\item{data}{dataframe}
}
\value{
new data without NA
}
\description{
Title relatively small in number then ignore those rows from the analysis
}
\examples{
omit_na(sample_read())
}
