% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/over_times_column.R
\name{over_times_column}
\alias{over_times_column}
\title{Title check the column's value is not above ~10 times the highest value ever seen in the historical data}
\usage{
over_times_column(hdata, data, colName)
}
\arguments{
\item{hdata}{historic data dataframe}

\item{data}{dataframe}

\item{colName}{column name to check in for the over range}
}
\value{
total rows that are over the given range
}
\description{
Title check the column's value is not above ~10 times the highest value ever seen in the historical data
}
\examples{
hnobleGasses <- data.frame(Gas=c("Helium","Neon", "Argon", "Krypton", "Xenon", "Radon", "Oganesson"),
Symbol=c("He","Ne","Ae","Kr","Xe","Rn","Og"), Nr=c(2,10,18,36,54,86,118),
Mass=c(4.003,20.18,39.948,83.8,131.293,222,294))
nobleGasses <- data.frame(Gas=c("Helium","Neon", "Argon", "Krypton", "Xenon", "Radon", "Oganesson"),
Symbol=c("He","Ne","Ae","Kr","Xe","Rn","Og"), Nr=c(2,10,18,36,54,86,118),
Mass=c(4.003,20.18,39.948,83.8,131.293,222,29400))
over_times_column(hnobleGasses, nobleGasses, "Mass")   # reutrns 1
}
