% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_model.R
\name{predict_model}
\alias{predict_model}
\title{Title}
\usage{
predict_model(model, data, intype)
}
\arguments{
\item{data}{test data from whole data}

\item{intype}{type of prediction "class", "response"}
}
\description{
Title
}
