% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printer.R
\name{printer}
\alias{printer}
\title{Title}
\usage{
printer(x, y)
}
\arguments{
\item{x}{what you want in the first column}

\item{y}{what you want in the second column}
}
\value{
A tibble
}
\description{
Title
}
\examples{
printer(x=rnorm(5), y=rnorm(5))
}
