% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processed_data.R
\name{processed_data}
\alias{processed_data}
\title{Title processed the read data by implementing all needed function}
\usage{
processed_data(data)
}
\arguments{
\item{data}{dataframe read}
}
\value{
processed/clean data
}
\description{
Title processed the read data by implementing all needed function
}
\examples{

}
