% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_read.R
\name{sample_read}
\alias{sample_read}
\title{Title Read data file (csv) from directory by the user}
\usage{
sample_read()
}
\value{
dataframe
}
\description{
Title Read data file (csv) from directory by the user
}
\examples{
sample_read()
}
