% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum_na.R
\name{sum_na}
\alias{sum_na}
\title{Count NAs in a vector}
\usage{
sum_na(x)
}
\arguments{
\item{x}{A vector}
}
\value{
number of NAs in x
}
\description{
Count NAs in a vector
}
\examples{
\dontrun{
   sumNa(airquality$Ozone)
   }
}
