% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_model_testdata.R
\name{test_model_testdata}
\alias{test_model_testdata}
\title{Title test the model with test dataset}
\usage{
test_model_testdata(data, divRatio, colName, intype)
}
\arguments{
\item{data}{dataframe}

\item{divRatio}{75, 80 % to divide the data in test and train}

\item{colName}{labeled column name}

\item{intype}{response, class}
}
\value{
confusion matrix table
}
\description{
Title test the model with test dataset
}
\examples{
test_model_testdata(data,divRatio,churn,"response")
}
