% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/under_column_range.R
\name{under_column_range}
\alias{under_column_range}
\title{Title check the column's that have value over the column range}
\usage{
under_column_range(data, colName, loVal)
}
\arguments{
\item{data}{dataframe}

\item{colName}{column name to check in for the over range}

\item{minVal}{lowest value of the range}
}
\value{
total rows that are under the given range
}
\description{
Title check the column's that have value over the column range
}
\examples{
nobleGasses <- data.frame(Gas=c("Helium","Neon", "Argon", "Krypton", "Xenon", "Radon", "Oganesson"),
Symbol=c("He","Ne","Ae","Kr","Xe","Rn","Og"), Nr=c(2,10,18,36,54,86,118),
Mass=c(4.003,20.18,39.948,83.8,131.293,222,294))
under_column_range(nobleGasses, "Nr", 5)   # reutrns 1
}
