% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unique_values_column.R
\name{unique_values_column}
\alias{unique_values_column}
\title{Title count unique values in a given column}
\usage{
unique_values_column(data, colName)
}
\arguments{
\item{data}{dataframe}

\item{colName}{column name}
}
\value{
unique values in a column
}
\description{
Title count unique values in a given column
}
