% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unique_values_column_2data.R
\name{unique_values_column_2data}
\alias{unique_values_column_2data}
\title{Title match column unique values of 2 dataframes}
\usage{
unique_values_column_2data(data1, data2, colName)
}
\arguments{
\item{data1}{dataframe}

\item{data2}{dataframe}

\item{colName}{column name}
}
\value{
TRUE if column have same unique values in two datasets
}
\description{
Title match column unique values of 2 dataframes
}
\examples{
#' nobleGasses <- data.frame(Gas=c("Helium","Neon", "Argon", "Krypton", "Xenon", "Radon", "Oganesson"),
Symbol=c("He","Ne","Ae","Kr","Xe","Rn","Og"), Nr=c(2,10,18,36,54,86,118),
Mass=c(4.003,20.18,39.948,83.8,131.293,222,294), UniqTest=c(2,3,4,2,4,3,5))

NnobleGasses <- data.frame(Gas=c("Helium","Neon", "Argon", "Krypton", "Xenon", "Radon", "Oganesson"),
Symbol=c("He","Ne","Ae","Kr","Xe","Rn","Og"), Nr=c(2,10,18,36,54,86,118),
Mass=c(4.003,20.18,39.948,83.8,222,222,222))

unique_values_column_2data(nobleGasses,NnobleGasses,"Gas")          #reutrn TRUE
unique_values_column_2data(nobleGasses,NnobleGasses,"Mass")          #reutrn FALSE
}
