% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_read.R
\name{sample_read}
\alias{sample_read}
\title{Read csv duplicate}
\usage{
sample_read(path)
}
\arguments{
\item{path}{path of filename}
}
\value{
a \code{tibble}
}
\description{
Read csv duplicate
}
\examples{
csv = system.file("extdata", "Telco_Customer_Churn.csv", package = "customerChurnUU")
sample_read(csv)
}
