% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unique_values_column.R
\name{unique_values_column}
\alias{unique_values_column}
\title{Title count unique values in a given column}
\usage{
unique_values_column(data, colName)
}
\arguments{
\item{data}{dataframe}

\item{colName}{column name}
}
\value{
unique values in a column
}
\description{
Title count unique values in a given column
}
\examples{
#' nobleGasses <- data.frame(Gas=c("Helium","Neon", "Argon", "Krypton", "Xenon", "Radon", "Oganesson"),
Symbol=c("He","Ne","Ae","Kr","Xe","Rn","Og"), Nr=c(2,10,18,36,54,86,118),
Mass=c(4.003,20.18,39.948,83.8,131.293,222,294), UniqTest=c(2,3,4,2,4,3,5))
unique_values_column(nobleGasses,"Gas")          #reutrn 7
unique_values_column(nobleGasses,"UniqTest")     # retrun 4
}
